+++ examples/event-processor-bleeps/CHANGELOG.md
@@ -1,5 +1,11 @@
 # event-processor-bleeps
 
+## 0.0.58
+
+### Patch Changes
+
+- ethereum-indexer-js-processor@0.7.1
+
 ## 0.0.57
 
 ### Patch Changes