+++ examples/mud/CHANGELOG.md
@@ -1,5 +1,18 @@
 # web-demo
 
+## 0.1.71
+
+### Patch Changes
+
+- stream option: doNotFetchUnfinalizedLogs
+- Updated dependencies
+  - ethereum-indexer@0.6.22
+  - ethereum-indexer-browser@0.7.8
+  - ethereum-indexer-js-processor@0.7.1
+  - event-processor-bleeps@0.0.58
+  - event-processor-conquest-eth@0.0.58
+  - event-processor-nfts@0.0.58
+
 ## 0.1.70
 
 ### Patch Changes